requirements.txt
SQLAlchemy~=2.0.32
python-dotenv~=1.0.1
requests~=2.32.3
aiogram~=3.10.0
httpx~=0.27.0
groq~=0.9.0
openai~=1.40.1
pydantic~=2.8.2
sqlalchemy-cockroachdb
asyncpg
redis
httpx[socks]